+++ optimization_model.h
@@ -4,7 +4,7 @@
 #include "data_structures.h"
 #include "err.h"
 
-errno_t gradient_descent(factors_t *sp, const int num_iters, const float lrn_rate, const data_samples_t *d_trn);
+errno_t gradient_descent(factors_t *sp, const int num_iters, const double lrn_rate, const data_samples_t *d_trn);
 
 errno_t validate_model(const data_input_t *din, data_output_t *dout);
